Distribution tests, improved model
Now both response parameters (counts and theta-hat) are tested for goodness-of-fit to several distributions (poisson, gaussian and negative binomial). The new per-cell multinomial model is built with those tests in mind. Performance metrics are calculated for each model and summarized to give a view of overall fit to the multi-cell data.